Replaces home page with chat
